+++ gnubg/renderprefs.c
@@ -113,7 +113,7 @@ SetColourSpeckle(const char *sz, unsigned char anColour[], int *pnSpeckle)
 
 /* Set colour (with floats) */
 static int
-SetColourX(gdouble arColour[4], const char *sz)
+SetColourX(float arColour[4], const char *sz)
 {
 
     char *pch;
@@ -244,7 +244,7 @@ SetMaterialDice(Material * pMat, const char *sz, int *flag)
 
 /* Set colour, alpha, refraction, shine, specular. */
 static int
-SetColourARSS(double aarColour[2][4],
+SetColourARSS(float aarColour[2][4],
               float arRefraction[2], float arCoefficient[2], float arExponent[2], char *sz, int i)
 {
 
@@ -257,7 +257,7 @@ SetColourARSS(double aarColour[2][4],
 
         if (pch) {
             /* alpha */
-            aarColour[i][3] = g_ascii_strtod(pch, NULL);
+            aarColour[i][3] = (float) g_ascii_strtod(pch, NULL);
 
             if ((pch = strchr(pch, ';')))
                 *pch++ = 0;
@@ -299,7 +299,7 @@ SetColourARSS(double aarColour[2][4],
 
 /* Set colour, shine, specular, flag. */
 static int
-SetColourSSF(gdouble aarColour[2][4],
+SetColourSSF(float aarColour[2][4],
              gfloat arCoefficient[2], gfloat arExponent[2], int afDieColour[2], char *sz, int i)
 {
